feat: Refocus editor after the terminal closes (#14)
### Description

This pull request improves the UX for the terminal launched by focus
back to editor group when terminal is closed

### Problem

When a terminal associated with a generic channel task is closed (either
manually by the user or upon task completion), the editor focus might
remain on the (now potentially empty) terminal panel or become unclear.
If the channel handler is configured to open related files upon closure
(e.g., log files), the user has to manually click back into the editor
group to view them.

### Changes proposed in this PR
1. Added a call to focus active editor group in listener to ensures that
immediately after the terminal is closed, the focus shifts back to the
active editor group.
2.  Added an explicit type to fix `any` type